ddt: split internal DDT API into separate header
Just to make it easier to know which bits to pay attention to.

Sponsored-by: Klara, Inc.
Sponsored-by: iXsystems, Inc.
